refactor: minor cleanup
